fastapi-backend/pyproject.toml
[tool.poetry]
name = "fastapi-backend"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.2"
pydantic = "^2.9.2"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = "^3.3.0"
sqlalchemy = "^2.0.36"
psycopg2-binary = "^2.9.10"
uvicorn = {extras = ["standard"], version = "^0.32.0"}
python-multipart = "^0.0.12"
redis = "^5.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
